+++ CHANGELOG.md
@@ -13,6 +13,11 @@ If your change does not need a CHANGELOG entry, add the "skip changelog" label t
 
 ## Unreleased
 
+### Breaking Changes
+
+- refactor: remove BedrockRuntime, SecretsManager, StepFunction, SNS instrumentation patches.
+  ([#446](https://github.com/aws-observability/aws-otel-js-instrumentation/pull/446))
+
 ## v0.11.0 - 2026-04-30
 
 - Support environment-configured endpoint visibility for HTTP operation names